attendance/requirements.txt
asgiref==3.8.1
Django==5.1.5
mysqlclient==2.2.7
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
